requirements.txt
datasets== 2.7.1
gql==3.4.0
matplotlib==3.4.3
pandas==1.5.1
requests==2.26.0
seaborn== 0.11.2
session_info==1.0.0
tqdm==4.64.1
openai==0.23.0
jsonlines==3.1.0
python-dotenv==0.21.0
brotli==1.0.9
brotlipy==0.7.0
ConfigParser==5.3.0
cryptography==38.0.3
Cython==0.29.24
dl==0.1.0
docutils==0.17.1
HTMLParser==0.0.2
importlib_metadata==4.8.1
keyring==23.1.0
lockfile==0.12.2
lxml==4.6.3
ordereddict==1.1
path.py==12.5.0
Pebble==4.6.3
protobuf==4.21.12
pyOpenSSL==23.0.0
toml==0.10.2
tornado==6.1
zipp==3.6.0
scipy==1.7.3